modified controller classes to allow easier mapping
